pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fx-regime-lab-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.40",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.4.0",
    "openai>=1.30.0",
    "fredapi>=0.5.1",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pytest>=8.1.0",
    "pre-commit>=3.7.0",
    "types-requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["src/legacy"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.ruff.lint.per-file-ignores]
"seed_history.py" = ["E402"]
"seed_macro_history.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.legacy.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
